+++ packages/stylist-core/src/ast/rule_content.rs
@@ -26,7 +26,7 @@ impl From<ScopeContent> for RuleContent {
 }
 
 impl ToStyleStr for RuleContent {
-    fn write_style<W: fmt::Write>(&self, w: &mut W, class_name: &str) -> fmt::Result {
+    fn write_style<W: fmt::Write>(&self, w: &mut W, class_name: Option<&str>) -> fmt::Result {
         match self {
             RuleContent::Block(ref b) => b.write_style(w, class_name),
             RuleContent::Rule(ref r) => r.write_style(w, class_name),