+++ packages/stylist-core/src/ast/to_style_str.rs
@@ -3,12 +3,14 @@ use std::fmt;
 /// Structs implementing this trait should be able to turn into
 /// a part of a CSS style sheet.
 pub trait ToStyleStr {
-    fn to_style_str(&self, class_name: &str) -> String {
+    fn to_style_str(&self, class_name: Option<&str>) -> String {
         let mut s = String::new();
 
         self.write_style(&mut s, class_name).unwrap();
 
         s
     }
-    fn write_style<W: fmt::Write>(&self, w: &mut W, class_name: &str) -> fmt::Result;
+
+    // If None is passed as class_name, it means to write a global style.
+    fn write_style<W: fmt::Write>(&self, w: &mut W, class_name: Option<&str>) -> fmt::Result;
 }