+++ packages/stylist/src/bindings/mod.rs
@@ -1,5 +1,0 @@
-// Copyright © 2020 Lukas Wagner
-
-#[cfg_attr(documenting, doc(cfg(feature = "yew_integration")))]
-#[cfg(feature = "yew_integration")]
-pub mod yew;